Create 3504343
Hi Damion and others who may potentially see this file! So far I have had very little experience with creating and writing code, but as I have been investigating the functions on this GIThub page, I am able to see the functionality and importance in learning code! The Midterm HTML web page was a dip in the water compared to the functions that we can learn and mess around with on this GIThub organization site. I think I am starting to grow to be extremely fond of coding and coding related technology skills! In terms of your coding skills, how involved have you been with coding previous to this course? Do you find yourself to be proficient, adjusted, or completely new to coding? Thank you so much for considering and making the effort to form this partnership, and being considerate of my times available to work! I hope that I am able to use this page and create further repositories in the future, and be self driven when it comes to digital creation from coding, design, and marketing!